Create Maven branch
