+++ jpushbullet/src/main/java/com/shakethat/jpushbullet/net/Devices.java
@@ -1,0 +1,29 @@
+
+package com.shakethat.jpushbullet.net;
+
+import java.util.List;
+
+public class Devices{
+   	private Extras extras;
+   	private Number id;
+   	private String iden;
+
+ 	public Extras getExtras(){
+		return this.extras;
+	}
+	public void setExtras(Extras extras){
+		this.extras = extras;
+	}
+ 	public Number getId(){
+		return this.id;
+	}
+	public void setId(Number id){
+		this.id = id;
+	}
+ 	public String getIden(){
+		return this.iden;
+	}
+	public void setIden(String iden){
+		this.iden = iden;
+	}
+}
